dockyard_backend/requirements.txt
fastapi
uvicorn
pydantic-settings
google-generativeai
supabase
httpx

